el9 and f37 support added

+++ bin/build-rpm.sh
@@ -48,31 +48,43 @@ clean(){
    # Tidy .pyc files
    find $APPRISE_DIR -name '*.pyc' -delete &>/dev/null
    find $APPRISE_DIR -type d -name '__pycache__' -exec rm -rf {} \ &>/dev/null;
-   # Remove previous `dist` and egg-info directory if it exists
-   [ -d $APPRISE_DIR/dist ] && rm -rf $APPRISE_DIR/dist
-   [ -d $APPRISE_DIR/apprise.egg-info ] && rm -rf $APPRISE_DIR/apprise.egg-info
-   [ -d $APPRISE_DIR/build ] && rm -rf $APPRISE_DIR/build
+   # Remove previously build details
+   [ -d "$APPRISE_DIR/apprise.egg-info" ] && rm -rf $APPRISE_DIR/apprise.egg-info
+   [ -d "$APPRISE_DIR/build" ] && rm -rf $APPRISE_DIR/build
+   [ -d "$APPRISE_DIR/BUILDROOT" ] && rm -rf $APPRISE_DIR/BUILDROOT
 }
 
 build(){
-   # Build Apprise
-   if [ ! -x $HOME/dev/bin/activate ]; then
-      $VENV_CMD $HOME/dev
-      [ $? -ne 0 ] && return 1
-   fi
-   . $HOME/dev/bin/activate
-   $PIP install coverage babel wheel markdown
-
-   pushd $APPRISE_DIR
-   # Build Man Page
-   ronn --roff $APPRISE_DIR/packaging/man/apprise.md
-   $PYTHON setup.py extract_messages
-   $PYTHON setup.py sdist
-
-   # exit from our virtual environment
-   deactivate
+   # Test spec file for any issues
+   rpmlint "$APPRISE_DIR/packaging/redhat/python-apprise.spec"
+   [ $? -ne 0 ] && echo "RPMLint Failed!" && return 1
 
    # Prepare RPM Package
+   # Detect our version
+   local VER=$(rpmspec -q --qf "%{version}\n" \
+      "$APPRISE_DIR/packaging/redhat/python-apprise.spec" 2>/dev/null | head -n1)
+   [ -z "$VER" ] && echo "Could not detect Apprise RPM Version" && return 1
+
+   if [ ! -f "$APPRISE_DIR/dist/apprise-$VER.tar.gz" ]; then
+      # Build Apprise
+      if [ ! -x $HOME/dev/bin/activate ]; then
+         $VENV_CMD $HOME/dev
+         [ $? -ne 0 ] && echo "Could not create Virtual Python Environment" && return 1
+      fi
+      . $HOME/dev/bin/activate
+      $PIP install coverage babel wheel markdown
+
+      pushd $APPRISE_DIR
+      # Build Man Page
+      ronn --roff $APPRISE_DIR/packaging/man/apprise.md
+      $PYTHON setup.py extract_messages
+      $PYTHON setup.py sdist
+
+      # exit from our virtual environment
+      deactivate
+   fi
+
+   # Prepare our RPM Source and SPEC dependencies
    find "$APPRISE_DIR/packaging/man/" -type f -name '*.1' \
          -exec cp --verbose {} "$APPRISE_DIR/dist" \;
    find "$APPRISE_DIR/packaging/redhat" -type f -name '*.patch' \
@@ -80,9 +92,9 @@ build(){
    find "$APPRISE_DIR/packaging/redhat" -type f -name '*.spec' \
          -exec cp --verbose {} "$APPRISE_DIR/dist" \;
 
+   # Build and Test our RPM Package
    rpmbuild -ba "$APPRISE_DIR/dist/python-apprise.spec"
-
-   return 0
+   return $?
 }
 
 # Prepare our environment
@@ -93,3 +105,6 @@ clean
 
 # Build
 build
+
+# Return our build status
+exit $?
